sif: Move IrqPin, system_irq and dt_irq out of the PCI subsystem into irq.rs
